requirements.txt
discord.py==2.3.2
pyyaml==6.0.1
py-dactyl==2.0.4
mcstatus==11.1.1